fix(025): 90*86400000 estourava int4 — 'integer out of range' derrubava a RPC na 1a exclusao real com tombstone (achado no E2E); forca aritmetica numeric

+++ backend/migrations/025_merge_por_elemento_tombstones.sql
@@ -110,8 +110,10 @@ as $fn$
     ) t
     group by t.key
   ) m
+  -- 90 dias em ms como NUMERIC: `90 * 86400000` em int4 estoura (integer out of
+  -- range) e derrubava a RPC inteira na 1ª exclusão real (achado no E2E 10/07).
   where m.maxts is not null
-    and m.maxts > extract(epoch from now()) * 1000 - 90 * 86400000
+    and m.maxts > extract(epoch from now()) * 1000 - (90::numeric * 86400000)
 $fn$;
 
 -- União do objeto _tombs inteiro ({diary:{...}, exercises:{...}, ...}).
